improved auth test
